feat: implement OTP email verification with DI setup and refactor aut…

+++ furniforge-backend/src/infrastructure/external-services/OtpService.ts
@@ -1,13 +1,19 @@
 import { IOtpService } from "@domain/services/IOtpservice.js";
 import { IOTPRepository } from "@domain/repositories/IOTPRepository.js";
 import { OtpToken } from "@domain/entities/OtpToken.js";
-import { ConflictError, BadRequestError } from "@domain/errors/AppError.js";
+import { ConflictError, BadRequestError, TooManyRequestsError } from "@domain/errors/AppError.js";
 import { ERROR_MESSAGES } from "@infrastructure/config/messages.js";
+import { env } from "@infrastructure/config/env.js";
+import { injectable, inject } from "inversify";
+import { TYPES } from "@infrastructure/di/types.js";
 
+@injectable()
 export class OtpService implements IOtpService {
-  private readonly TTL = 300;
+  private readonly TTL = env.OTP.EXPIRY;
 
-  constructor(private otpRepository: IOTPRepository) {}
+  constructor(
+    @inject(TYPES.IOTPRepository) private otpRepository: IOTPRepository
+  ) {}
 
   generateOTP(): string {
     return Math.floor(100000 + Math.random() * 900000).toString();
@@ -30,31 +36,47 @@ export class OtpService implements IOtpService {
 
     await this.otpRepository.save(otpToken, this.TTL);
 
+    console.log("Saved OTP:", otpToken);
+    console.log("User OTP Key:", `otp:user:${userId}`);
+    console.log("User OTP Key:", `otp:${otpToken.otpId}`);
+    console.log("Code Key:", `otp:code:${email}:${otpCode}`);
+    console.log("TTL:", `otp:user:${userId}`);
+
     return otpToken;
   }
 
   async verifyOtp( userId: string, email: string, inputOtp: string ): Promise<OtpToken> {
 
-    const existing = await this.otpRepository.getByCode(inputOtp, email);
+    const existing = await this.otpRepository.getByUserId(userId);
 
     if (!existing) {
       throw new BadRequestError(ERROR_MESSAGES.AUTH.OTP_NOT_FOUND);
     }
 
+    if (existing.email !== email) {
+      throw new BadRequestError("Invalid OTP request");
+    }
+
     if (existing.isExpired()) {
-      await this.otpRepository.delete(existing);
       throw new BadRequestError(ERROR_MESSAGES.AUTH.OTP_EXPIRED);
     }
 
     try {
       existing.verify(inputOtp);
     } catch (err: any) {
-      await this.otpRepository.update(existing); 
-      throw new BadRequestError(err.message);
+
+      await this.otpRepository.update(existing);
+
+      if(err instanceof TooManyRequestsError){
+        await this.otpRepository.delete(existing)
+      }
+      err.meta = {
+        remainingAttempts: existing.remainingAttempts
+      }
+      throw err;
     }
-
+    
     await this.otpRepository.delete(existing);
-
     return existing;
   }
 }
